Store BigDecimal and BigInteger as Numbers like the Java SDK does. (#1933)
Also be able to read BigDecimal and BigInteger that were written as String.
Also does not lose precision by converting BigDecimal to double in the transcoder.

Closes #1611.

+++ src/main/java/org/springframework/data/couchbase/core/convert/translation/JacksonTranslationService.java
@@ -220,7 +220,7 @@ public class JacksonTranslationService implements TranslationService, Initializi
 			case VALUE_NUMBER_INT:
 				return parser.getNumberValue();
 			case VALUE_NUMBER_FLOAT:
-				return parser.getDoubleValue();
+				return parser.getDecimalValue();
 			case VALUE_NULL:
 				return null;
 			default:
